+++ src/gui/widgets/chord_object.c
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2018-2021 Alexandros Theodotou <alex@zrythm.org>
+// SPDX-FileCopyrightText: © 2018-2022 Alexandros Theodotou <alex@zrythm.org>
 // SPDX-License-Identifier: LicenseRef-ZrythmLicense
 
 #include "audio/chord_object.h"
@@ -60,11 +60,12 @@ chord_object_draw (ChordObject * self, GtkSnapshot * snapshot)
   /* create clip */
   GskRoundedRect  rounded_rect;
   graphene_rect_t graphene_rect = GRAPHENE_RECT_INIT (
-    obj->full_rect.x, obj->full_rect.y, obj->full_rect.width,
-    obj->full_rect.height);
+    (float) obj->full_rect.x, (float) obj->full_rect.y,
+    (float) obj->full_rect.width,
+    (float) obj->full_rect.height);
   gsk_rounded_rect_init_from_rect (
     &rounded_rect, &graphene_rect,
-    obj->full_rect.height / 6.0f);
+    (float) obj->full_rect.height / 6.0f);
   gtk_snapshot_push_rounded_clip (snapshot, &rounded_rect);
 
   /* fill */
@@ -89,8 +90,8 @@ chord_object_draw (ChordObject * self, GtkSnapshot * snapshot)
   gtk_snapshot_translate (
     snapshot,
     &GRAPHENE_POINT_INIT (
-      obj->full_rect.x + CHORD_OBJECT_NAME_PADDING,
-      obj->full_rect.y + CHORD_OBJECT_NAME_PADDING));
+      (float) obj->full_rect.x + CHORD_OBJECT_NAME_PADDING,
+      (float) obj->full_rect.y + CHORD_OBJECT_NAME_PADDING));
   pango_layout_set_text (self->layout, display_str, -1);
   gtk_snapshot_append_layout (snapshot, self->layout, &c2);
   gtk_snapshot_restore (snapshot);
